rename friendModal
